manager/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-socketio==5.11.2
pydantic==2.5.2
httpx==0.25.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
